+++ crates/forge/src/lib.rs
@@ -2,12 +2,14 @@ use crate::compatibility_check::{Requirement, RequirementsChecker, create_versio
 use anyhow::Result;
 use anyhow::anyhow;
 use camino::Utf8PathBuf;
-use clap::{Parser, Subcommand, ValueEnum};
+use clap::{CommandFactory, Parser, Subcommand, ValueEnum};
 use forge_runner::CACHE_DIR;
+use forge_runner::forge_config::ForgeTrackedResource;
 use run_tests::workspace::run_for_workspace;
 use scarb_api::{ScarbCommand, metadata::MetadataCommandExt};
 use scarb_ui::args::{FeaturesSpec, PackagesFilter};
 use semver::Version;
+use shared::auto_completions::{Completion, generate_completions};
 use shared::print::print_as_warning;
 use std::cell::RefCell;
 use std::ffi::OsString;
@@ -36,6 +38,7 @@ const MINIMAL_SCARB_VERSION: Version = Version::new(2, 7, 0);
 const MINIMAL_RECOMMENDED_SCARB_VERSION: Version = Version::new(2, 8, 5);
 const MINIMAL_SCARB_VERSION_PREBUILT_PLUGIN: Version = Version::new(2, 10, 0);
 const MINIMAL_USC_VERSION: Version = Version::new(2, 0, 0);
+const MINIMAL_SCARB_VERSION_FOR_SIERRA_GAS: Version = Version::new(2, 10, 0);
 
 #[derive(Parser, Debug)]
 #[command(
@@ -100,6 +103,8 @@ enum ForgeSubcommand {
     CleanCache {},
     /// Check if all `snforge` requirements are installed
     CheckRequirements,
+    /// Generate completion script
+    Completion(Completion),
 }
 
 #[derive(Parser, Debug)]
@@ -195,6 +200,10 @@ pub struct TestArgs {
     #[arg(long)]
     no_optimization: bool,
 
+    /// Specify tracked resource type
+    #[arg(long, value_enum, default_value_t)]
+    tracked_resource: ForgeTrackedResource,
+
     /// Additional arguments for cairo-coverage or cairo-profiler
     #[clap(last = true)]
     additional_args: Vec<OsString>,
@@ -250,7 +259,7 @@ pub fn main_execution() -> Result<ExitStatus> {
             Ok(ExitStatus::Success)
         }
         ForgeSubcommand::Test { args } => {
-            check_requirements(false)?;
+            check_requirements(false, args.tracked_resource)?;
             let cores = if let Ok(available_cores) = available_parallelism() {
                 available_cores.get()
             } else {
@@ -266,23 +275,51 @@ pub fn main_execution() -> Result<ExitStatus> {
             rt.block_on(run_for_workspace(args))
         }
         ForgeSubcommand::CheckRequirements => {
-            check_requirements(true)?;
+            check_requirements(true, ForgeTrackedResource::default())?;
+            Ok(ExitStatus::Success)
+        }
+        ForgeSubcommand::Completion(completion) => {
+            generate_completions(completion.shell, &mut Cli::command())?;
             Ok(ExitStatus::Success)
         }
     }
 }
 
-fn check_requirements(output_on_success: bool) -> Result<()> {
+fn check_requirements(
+    output_on_success: bool,
+    forge_tracked_resource: ForgeTrackedResource,
+) -> Result<()> {
     let mut requirements_checker = RequirementsChecker::new(output_on_success);
-    requirements_checker.add_requirement(Requirement {
-        name: "Scarb".to_string(),
-        command: RefCell::new(ScarbCommand::new().arg("--version").command()),
-        minimal_version: MINIMAL_SCARB_VERSION,
-        minimal_recommended_version: Some(MINIMAL_RECOMMENDED_SCARB_VERSION),
-        helper_text: "Follow instructions from https://docs.swmansion.com/scarb/download.html"
-            .to_string(),
-        version_parser: create_version_parser("Scarb", r"scarb (?<version>[0-9]+.[0-9]+.[0-9]+)"),
-    });
+    match forge_tracked_resource {
+        ForgeTrackedResource::CairoSteps => {
+            requirements_checker.add_requirement(Requirement {
+                name: "Scarb".to_string(),
+                command: RefCell::new(ScarbCommand::new().arg("--version").command()),
+                minimal_version: MINIMAL_SCARB_VERSION,
+                minimal_recommended_version: Some(MINIMAL_RECOMMENDED_SCARB_VERSION),
+                helper_text:
+                    "Follow instructions from https://docs.swmansion.com/scarb/download.html"
+                        .to_string(),
+                version_parser: create_version_parser(
+                    "Scarb",
+                    r"scarb (?<version>[0-9]+.[0-9]+.[0-9]+)",
+                ),
+            });
+        }
+        ForgeTrackedResource::SierraGas => {
+            requirements_checker.add_requirement(Requirement {
+                name: "Scarb".to_string(),
+                command: RefCell::new(ScarbCommand::new().arg("--version").command()),
+                minimal_version: MINIMAL_SCARB_VERSION_FOR_SIERRA_GAS,
+                minimal_recommended_version: None,
+                helper_text: "This version of scarb is the minimum required in order to support sierra gas tracking \
+                (it comes with sierra >= 1.7.0 support)\n\
+                Follow instructions from https://docs.swmansion.com/scarb/download.html"
+                    .to_string(),
+                version_parser: create_version_parser("Scarb", r"scarb (?<version>[0-9]+.[0-9]+.[0-9]+)"),
+            });
+        }
+    };
     requirements_checker.add_requirement(Requirement {
         name: "Universal Sierra Compiler".to_string(),
         command: RefCell::new(UniversalSierraCompilerCommand::new().arg("--version").command()),